[Refactor]: Chapter05 5.2 복잡한 메소드 구현 로직을 새로운 클래스로 분리
HttpRequest의 요청 라인을 처리하는 processRequestLine() 메소드의 복잡도가 높다.
RequestLine이라는 이름으로 새로운 클래스를 추가하는 방식으로 리팩토링

issue #1
